docker command in tools
