+++ Dockerfile
@@ -42,17 +42,17 @@ RUN apt-get update \
 # the final image.
 
 RUN cd /tmp \
- && git clone https://github.com/open-s4c/genmc.git genmc9 \
+ && git clone --branch "v0.9" --depth 1\
+     https://github.com/open-s4c/genmc.git genmc9 \
  && cd genmc9 \
- && git checkout "v0.9" \
  && autoreconf --install \
  && ./configure --prefix=/usr/share/genmc9 \
  && make install -j8
 
 RUN cd /tmp \
- && git clone https://github.com/open-s4c/genmc.git genmc10 \
+ && git clone --branch "v0.10.1-a" --depth 1 \
+     https://github.com/open-s4c/genmc.git genmc10 \
  && cd genmc10 \
- && git checkout "v0.10.1-a" \
  && autoreconf --install \
  && ./configure --prefix=/usr/share/genmc10 \
  && make install -j8
